+++ packages/container/src/WebWidgetUpdateEvent.js
@@ -1,0 +1,14 @@
+/* global Event */
+
+export class WebWidgetUpdateEvent extends Event {
+  #value = null;
+
+  constructor(type, init = {}) {
+    super(type, init);
+    this.#value = init.value;
+  }
+
+  get value() {
+    return this.#value;
+  }
+}