+++ packages/container/src/index.js
@@ -1,2 +1,2 @@
-export { HTMLWebWidgetElement, bootstrap } from './HTMLWebWidgetElement.js';
-export { WebWidgetDependencies } from './WebWidgetDependencies.js';
+export * from './HTMLWebWidgetElement.js';
+export * from './WebWidgetUpdateEvent.js';