+++ packages/system-loader/test/test.widget.js
@@ -3,14 +3,15 @@
 System.register([], exports => ({
   execute() {
     exports('default', () => ({
-      async mount({ container }) {
+      async mount({ container, test }) {
+        test();
         setTimeout(() => {
           container.update({
             data: {
               lifecycle: 'mount'
             }
           });
-        });
+        }, 50);
       }
     }));
   }