+++ packages/umd-loader/test/test2.widget.js
@@ -1,5 +1,6 @@
 window.TestWidget2 = () => ({
-  async mount({ container }) {
+  async mount({ container, test }) {
+    test();
     setTimeout(() => {
       container.update({
         data: {